[Broken] More intermediate progress. Split up the Input devices.
