+++ Code/Engine/Engine.vcxproj
@@ -33,7 +33,9 @@
     <ClCompile Include="Input\BinaryReader.cpp" />
     <ClCompile Include="Input\BinaryWriter.cpp" />
     <ClCompile Include="Input\Console.cpp" />
-    <ClCompile Include="Input\InputDevices.cpp" />
+    <ClCompile Include="Input\InputDevices\InputDevice.cpp" />
+    <ClCompile Include="Input\InputDevices\KeyboardInputDevice.cpp" />
+    <ClCompile Include="Input\InputDevices\MouseInputDevice.cpp" />
     <ClCompile Include="Input\InputMap.cpp" />
     <ClCompile Include="Input\InputOutputUtils.cpp" />
     <ClCompile Include="Input\InputSystem.cpp" />
@@ -123,7 +125,9 @@
     <ClInclude Include="Input\BinaryReader.hpp" />
     <ClInclude Include="Input\BinaryWriter.hpp" />
     <ClInclude Include="Input\Console.hpp" />
-    <ClInclude Include="Input\InputDevices.hpp" />
+    <ClInclude Include="Input\InputDevices\InputDevice.hpp" />
+    <ClInclude Include="Input\InputDevices\KeyboardInputDevice.hpp" />
+    <ClInclude Include="Input\InputDevices\MouseInputDevice.hpp" />
     <ClInclude Include="Input\InputMap.hpp" />
     <ClInclude Include="Input\InputOutputUtils.hpp" />
     <ClInclude Include="Input\InputSystem.hpp" />
