+++ Code/Engine/Input/InputDevices/MouseInputDevice.cpp
@@ -1,0 +1,65 @@
+#include "Engine/Input/InputDevices/MouseInputDevice.hpp"
+
+#ifdef WIN32
+#define PLATFORM_WINDOWS
+#define WIN32_LEAN_AND_MEAN
+#include <Windows.h>
+#endif
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::SetButtonValue(unsigned char mouseButton, bool isDown)
+{
+    m_buttons[mouseButton].SetValue(isDown ? 1.0f : 0.0f);
+}
+
+//-----------------------------------------------------------------------------------
+InputValue* MouseInputDevice::FindButtonValue(unsigned char mouseButton)
+{
+    return &(m_buttons[mouseButton]);
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::SetDelta(const Vector2Int& cursorDelta)
+{
+    m_deltaPosition.SetValue(Vector2(cursorDelta.x, -cursorDelta.y).GetNorm());
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::HideMouseCursor()
+{
+    ShowCursor(FALSE);
+    m_isCursorVisible = false;
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::ShowMouseCursor()
+{
+    ShowCursor(TRUE);
+    m_isCursorVisible = true;
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::LockMouseCursor()
+{
+    m_captureCursor = true;
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::UnlockMouseCursor()
+{
+    m_captureCursor = false;
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::CaptureMouseCursor()
+{
+    HideMouseCursor();
+    LockMouseCursor();
+}
+
+//-----------------------------------------------------------------------------------
+void MouseInputDevice::ReleaseMouseCursor()
+{
+    ShowMouseCursor();
+    UnlockMouseCursor();
+}
