+++ Code/Engine/Input/InputMap.hpp
@@ -16,17 +16,17 @@ public:
     ~InputMap();
 
     //FUNCTIONS/////////////////////////////////////////////////////////////////////
-    InputValue* AddInputValue(std::string const &name);
-    InputValue* AddInputValue(std::string const &name, InputValue *other);
+    InputValue* MapInputValue(std::string const &name);
+    InputValue* MapInputValue(std::string const &name, InputValue *other);
     InputValue* FindInputValue(std::string const &name);
     float GetValue(std::string const &name);
     bool IsDown(const std::string& name);
     bool IsUp(const std::string& name);
     bool WasJustPressed(const std::string& name);
     bool WasJustReleased(const std::string& name);
-    InputAxis* AddInputAxis(const std::string &name);
-    InputAxis* AddInputAxis(const std::string& name, InputAxis* inputAxis);
-    InputAxis* AddInputAxis(const std::string& name, InputValue* positiveInput, InputValue* negativeInput);
+    InputAxis* MapInputAxis(const std::string &name);
+    InputAxis* MapInputAxis(const std::string& name, InputAxis* inputAxis);
+    InputAxis* MapInputAxis(const std::string& name, InputValue* positiveInput, InputValue* negativeInput);
     InputAxis* FindInputAxis(std::string const &name);
     Vector2 GetVector2(const std::string& xName, const std::string& yName);
     void Clear();
@@ -34,4 +34,5 @@ public:
     //MEMBER VARIABLES/////////////////////////////////////////////////////////////////////
     std::map<std::string, InputValue*> m_values;
     std::map<std::string, InputAxis*> m_axies;
+    std::map<std::string, InputVector2*> m_vectors;
 };
